🗑️ Remove deprecated BedrockAgent implementation, consolidating functionality into a single agent file for improved maintainability and clarity.
